♻️ M26FS0 code: restore module size budget
